+++ Arduino/RemoteBOY/LED.cpp
@@ -1,11 +1,11 @@
 #include "LED.h"
 #include <Arduino.h>
 
-static const int LED_TOTAL_ON_TIME = 5000; // stay on for 5 seconds
-static const int LED_BLINK_FREQUENCY = 500; // blink every 0.5 seconds
-static const int PIN_LED_2 = D3;
-static const int PIN_LED_1 = D4;
-static const int ledPins[2] = {PIN_LED_2, PIN_LED_1};
+static const uint32_t LED_TOTAL_ON_TIME = 5000; // stay on for 5 seconds
+static const uint32_t LED_BLINK_FREQUENCY = 500; // blink every 0.5 seconds
+static const uint8_t PIN_LED_2 = D3;
+static const uint8_t PIN_LED_1 = D4;
+static const uint8_t ledPins[2] = {PIN_LED_2, PIN_LED_1};
 
 LED::LED(void) {}
